+++ src/Documenter.jl
@@ -57,14 +57,16 @@ export Deps, makedocs, deploydocs, hide, external
 
 """
     makedocs(
-        root      = "<current-directory>",
+        root    = "<current-directory>",
         repo_root = dirname(root),
-        source    = "src",
-        build     = "build",
-        clean     = true,
-        doctest   = true,
-        modules   = Module[],
-        repo      = "",
+        source  = "src",
+        build   = "build",
+        clean   = true,
+        doctest = true,
+        modules = Module[],
+        repo    = "",
+        highlightsig = true,
+        sitename = "",
     )
 
 Combines markdown files and inline docstrings into an interlinked document.
@@ -96,7 +98,7 @@ The folder structure that [`makedocs`](@ref) expects looks like:
 this keyword does not need to be set. It is, for the most part, needed when repeatedly
 running `makedocs` from the Julia REPL like so:
 
-    julia> makedocs(root = joinpath(pathof(MyModule), "..", "..", "docs"))
+    julia> makedocs(root = joinpath(dirname(pathof(MyModule)), "..", "docs"))
 
 **`repo_root`** is the top-level directory of the repository, assumed to be `root`'s
 parent. You only need to set this if you are making use of [`external`](@ref).
@@ -153,6 +155,14 @@ For example if you are using GitLab.com, you could use
 ```julia
 makedocs(repo = \"https://gitlab.com/user/project/blob/{commit}{path}#{line}\")
 ```
+
+**`highlightsig`** enables or disables automatic syntax highlighting of leading, unlabeled
+code blocks in docstrings (as Julia code). For example, if your docstring begins with an
+indented code block containing the function signature, then that block would be highlighted
+as if it were a labeled Julia code block. No other code blocks are affected. This feature
+is enabled by default.
+
+**`sitename`** is displayed in the title bar and/or the navigation menu when applicable.
 
 # Experimental keywords
 
@@ -804,194 +814,7 @@ function getenv(regex::Regex)
     error("could not find key/iv pair.")
 end
 
-<<<<<<< HEAD
 const external_dir = "DOCUMENTER_EXTERNAL"
-=======
-export Travis
-
-"""
-Package functions for interacting with Travis.
-
-$(EXPORTS)
-"""
-module Travis
-
-using Compat, DocStringExtensions
-
-export genkeys
-
-import Base.LibGit2.GITHUB_REGEX
-
-
-"""
-$(SIGNATURES)
-
-Generate ssh keys for package `package` to automatically deploy docs from Travis to GitHub
-pages. `package` can be either the name of a package or a path. Providing a path allows keys
-to be generated for non-packages or packages that are not found in the Julia `LOAD_PATH`.
-Use the `remote` keyword to specify the user and repository values.
-
-This function requires the following command lines programs to be installed:
-
-- `which`
-- `git`
-- `travis`
-- `ssh-keygen`
-
-# Examples
-
-```jlcon
-julia> using Documenter
-
-julia> Travis.genkeys("MyPackageName")
-[ ... output ... ]
-
-julia> Travis.genkeys("MyPackageName", remote="organization")
-[ ... output ... ]
-
-julia> Travis.genkeys("/path/to/target/directory")
-[ ... output ... ]
-```
-"""
-function genkeys(package; remote="origin")
-    # Error checking. Do the required programs exist?
-    success(`which which`)      || error("'which' not found.")
-    success(`which git`)        || error("'git' not found.")
-    success(`which ssh-keygen`) || error("'ssh-keygen' not found.")
-
-    directory = "docs"
-    filename  = ".documenter"
-
-    path = isdir(package) ? package : Pkg.dir(package, directory)
-    isdir(path) || error("`$path` not found. Provide a package name or directory.")
-
-    cd(path) do
-        # Check for old '$filename.enc' and terminate.
-        isfile("$filename.enc") &&
-            error("$package already has an ssh key. Remove it and try again.")
-
-        # Are we in a git repo?
-        success(`git status`) || error("'Travis.genkey' only works with git repositories.")
-
-        # Find the GitHub repo org and name.
-        user, repo =
-            let r = readchomp(`git config --get remote.$remote.url`)
-                m = match(GITHUB_REGEX, r)
-                m === nothing && error("no remote repo named '$remote' found.")
-                m[2], m[3]
-            end
-
-        # Generate the ssh key pair.
-        success(`ssh-keygen -N "" -f $filename`) || error("failed to generated ssh key pair.")
-
-        # Prompt user to add public key to github then remove the public key.
-        let url = "https://github.com/$user/$repo/settings/keys"
-            info("add the public key below to $url with read/write access:")
-            println("\n", read("$filename.pub", String))
-            rm("$filename.pub")
-        end
-
-        # Base64 encode the private key and prompt user to add it to travis. The key is
-        # *not* encoded for the sake of security, but instead to make it easier to
-        # copy/paste it over to travis without having to worry about whitespace.
-        let url = "https://travis-ci.org/$user/$repo/settings"
-            info("add a secure environment variable named 'DOCUMENTER_KEY' to $url with value:")
-            println("\n", base64encode(read(".documenter", String)), "\n")
-            rm(filename)
-        end
-    end
-end
-
-end
-
-"""
-$(SIGNATURES)
-
-Creates a documentation stub for a package called `pkgname`. The location of
-the documentation is assumed to be `<package directory>/docs`, but this can
-be overriden with the keyword argument `dir`.
-
-It creates the following files
-
-```
-docs/
-    .gitignore
-    src/index.md
-    make.jl
-    mkdocs.yml
-```
-
-# Arguments
-
-**`pkgname`** is the name of the package (without `.jl`). It is used to
-determine the location of the documentation if `dir` is not provided.
-
-# Keywords
-
-**`dir`** defines the directory where the documentation will be generated.
-It defaults to `<package directory>/docs`. The directory must not exist.
-
-# Examples
-
-```jlcon
-julia> using Documenter
-
-julia> Documenter.generate("MyPackageName")
-[ ... output ... ]
-```
-"""
-function generate(pkgname::AbstractString; dir=nothing)
-    # TODO:
-    #   - set up deployment to `gh-pages`
-    #   - fetch url and username automatically (e.g from git remote.origin.url)
-
-    # Check the validity of the package name
-    if length(pkgname) == 0
-        error("Package name can not be an empty string.")
-    end
-    # Determine the root directory where we wish to generate the docs and
-    # check that it is a valid directory.
-    docroot = if dir === nothing
-        pkgdir = Pkg.dir(pkgname)
-        if !isdir(pkgdir)
-            error("Unable to find package $(pkgname).jl at $(pkgdir).")
-        end
-        joinpath(pkgdir, "docs")
-    else
-        dir
-    end
-
-    if ispath(docroot)
-        error("Directory $(docroot) already exists.")
-    end
-
-    # deploy the stub
-    try
-        info("Deploying documentation to $(docroot)")
-        mkdir(docroot)
-
-        # create the root doc files
-        Generator.savefile(docroot, ".gitignore") do io
-            write(io, Generator.gitignore())
-        end
-        Generator.savefile(docroot, "make.jl") do io
-            write(io, Generator.make(pkgname))
-        end
-        Generator.savefile(docroot, "mkdocs.yml") do io
-            write(io, Generator.mkdocs(pkgname))
-        end
-
-        # Create the default documentation source files
-        Generator.savefile(docroot, "src/index.md") do io
-            write(io, Generator.index(pkgname))
-        end
-    catch
-        rm(docroot, recursive=true)
-        rethrow()
-    end
-    nothing
-end
->>>>>>> Rework external pages
 
 """
     external(path; target) -> ExternalPage