+++ src/Writers/HTMLWriter.jl
@@ -228,6 +228,7 @@ getpage(ctx, navnode::Documents.NavNode) = getpage(ctx, navnode.page)
 
 
 function render(doc::Documents.Document, settings::HTML=HTML())
+    @info "HTMLWriter: rendering HTML pages."
     !isempty(doc.user.sitename) || error("HTML output requires `sitename`.")
 
     ctx = HTMLContext(doc, settings)
@@ -443,7 +444,8 @@ function render_navmenu(ctx, navnode)
     navmenu = nav[".toc"]
     if !isempty(ctx.logo)
         push!(navmenu.nodes,
-            a[:href => relhref(src, "index.html")](
+            # the logo will point to the first page in the navigation menu
+            a[:href => navhref(ctx, first(ctx.doc.internal.navlist), navnode)](
                 img[
                     ".logo",
                     :src => relhref(src, ctx.logo),
@@ -914,8 +916,8 @@ Returns the full path of a [`Documents.NavNode`](@ref) relative to `src/`.
 get_url(ctx, navnode::Documents.NavNode) = get_url(ctx, navnode.page)
 
 """
-If `html_prettyurls` is enabled, returns a "pretty" version of the `path` which can then be
-used in links in the resulting HTML file.
+If `prettyurls` for [`HTML`](@ref Documenter.HTML) is enabled, returns a "pretty" version of
+the `path` which can then be used in links in the resulting HTML file.
 """
 function pretty_url(ctx, path::AbstractString)
     if ctx.settings.prettyurls
@@ -1053,10 +1055,28 @@ function mdconvert(paragraph::Markdown.Paragraph, parent::Markdown.List; kwargs.
     return (list_has_loose_field && !parent.loose) ? content : Tag(:p)(content)
 end
 
-mdconvert(t::Markdown.Table, parent; kwargs...) = Tag(:table)(
-    Tag(:tr)(map(x -> Tag(:th)(mdconvert(x, t; kwargs...)), t.rows[1])),
-    map(x -> Tag(:tr)(map(y -> Tag(:td)(mdconvert(y, x; kwargs...)), x)), t.rows[2:end])
-)
+function mdconvert(t::Markdown.Table, parent; kwargs...)
+    @tags table tr th td
+    alignment_style = map(t.align) do align
+        if align == :r
+            "text-align: right"
+        elseif align == :c
+            "text-align: center"
+        else
+            "text-align: left"
+        end
+    end
+    table(
+        tr(map(enumerate(t.rows[1])) do (i, x)
+            th[:style => alignment_style[i]](mdconvert(x, t; kwargs...))
+        end),
+        map(t.rows[2:end]) do x
+            tr(map(enumerate(x)) do (i, y) # each cell in a row
+                td[:style => alignment_style[i]](mdconvert(y, x; kwargs...))
+            end)
+        end
+    )
+end
 
 mdconvert(expr::Union{Expr,Symbol}, parent; kwargs...) = string(expr)