+++ src/Writers/LaTeXWriter.jl
@@ -4,7 +4,7 @@ A module for rendering `Document` objects to LaTeX and PDF.
 # Keywords
 
 [`LaTeXWriter`](@ref) uses the following additional keyword arguments that can be passed to
-[`Documenter.makedocs`](@ref): `authors`, `sitename`.
+[`makedocs`](@ref Documenter.makedocs): `authors`, `sitename`.
 
 **`sitename`** is the site's title displayed in the title bar and at the top of the
 navigation menu. It goes into the `\\title` LaTeX command.
@@ -19,7 +19,7 @@ import ...Documenter: Documenter
     LaTeXWriter.LaTeX(; kwargs...)
 
 Output format specifier that results in LaTeX/PDF output.
-Used together with [`makedocs`](@ref), e.g.
+Used together with [`makedocs`](@ref Documenter.makedocs), e.g.
 
 ```julia
 makedocs(
@@ -85,6 +85,7 @@ const DOCUMENT_STRUCTURE = (
 )
 
 function render(doc::Documents.Document, settings::LaTeX=LaTeX())
+    @info "LaTeXWriter: rendering PDF."
     mktempdir() do path
         cp(joinpath(doc.user.root, doc.user.build), joinpath(path, "build"))
         cd(joinpath(path, "build")) do
@@ -534,6 +535,7 @@ function latexinline(io::IO, md::Markdown.Image)
         else
             normpath(joinpath(dirname(io.filename), md.url))
         end
+        url = replace(url, "\\" => "/") # use / on Windows too.
         wrapinline(io, "includegraphics") do
             _print(io, url)
         end